requirements.txt
asgiref==3.8.1
Django==4.2.17
djangorestframework==3.12.4
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
web3==6.20.1
